requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
databases==0.2.5
Django==2.2.3
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-request==1.5.4
django-tinymce-lite==0.0.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.16
numpy==1.16.4
pbr==5.4.0
Pillow==6.1.0
protobuf==3.9.0
pylint==2.3.1
pysqlite3==0.2.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
SQLite3Database==0.2.0
sqlparse==0.3.0
stevedore==1.30.1
stripe==2.35.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.6.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
